requirements.txt
fastapi
uvicorn
python-multipart
chromadb
langchain
langchain-community
pypdf
python-docx
python-dotenv
httpx
requests
groq
